+++ ios/Classes/Models/Ledger/Rekey/RekeyTransactionSendDraft.swift
@@ -20,8 +20,8 @@ import Foundation
 struct RekeyTransactionSendDraft: TransactionSendDraft {
     var from: Account
     var toAccount: String?
-    var amount: Double?
-    var fee: Int64?
+    var amount: Decimal?
+    var fee: UInt64?
     var isMaxTransaction = false
     var identifier: String?
     var note: String?
